animations-dbus: Add client library
The client library more or less thinly wraps the server library
and supports both async and sync operations. The sync operations
wrap the async operations in a pushed main context.

The benefit of using the client library is that there isn't the
extra step invovled in getting the object path and then creating
a proxy for the object, it can be done in one step.