animations-dbus: Add server side animations-dbus library
The server side library is designed to be used like GNOME-Shell
or other compositors which implement surface level animations.

The server uses animations_dbus_server_new_async() to asynchronously
create a new AnimationsDbusServer object and corresponding
AnimationsDbusConnectionManager which will be exported on the bus. One
of the construct properties that the server must provide is an implementation
of a AnimationsDbusServerEffectFactory which is responsible for constructing
the effect "metaclass" (AnimationsDbusServerEffectBridge) on the server side.

On each surface that the server is managing, the server calls
animations_dbus_server_register_surface along with an
implementation of a AnimationsDbusServerSurfaceBridge, which
is an interface providing properties about the surface that
would be interesting to internal clients. Calling register_surface
will export an AnimationsDbusSurface object on the bus for clients
to interact with.

The AnimationsDbusServerSurfaceBridge implementation should have
an attach_effect method which takes an AnimationsDbusServerEffectBridge
and determines if it can be attached to the surface on that event. If
it can be, attach_effect should return an AnimationsDbusServerSurfaceAttachedEffect
implementation, behind which will be private data for that animation
(for instance, progress, spring state) that exists for that animation
on the given surface. It is the server's responsibility to actually
perform the animation using that private data and the settings for that
animation exist on the attached AnimationsDbusServerEffectBridge.

Clients create and attach animations by first registering themselves
using the RegisterClient method on the exported AnimationsDbusConnectionManager
which takes note of their bus name and returns an object path to an
AnimationsDbusAnimationManager (specific to that client), which they can
call CreateAnimation on to create new animation metaclasses. Clients can
then call the AttachEffect method on the exported AnimationsDbusSurface
to attach their effects to the surface. Once the bus name owned by the
client disappears from the bus, all the client specific state will
be automatically cleaned up (eg, animations are detached, private data
for attached animations is destroyed).

Surfaces can have multiple animations attached to a single event. In
case multiple animations are attached, the highest priority attached one
is selected. If the highest priority animation is destroyed, it is removed
from the list and the next highest priority animation will be used. In
general, the latest attached effect by a client has the highest priority.

Both external clients and the server can register animations with
the server (the server would create a new AnimationManager object
using the animation_dbus_server_create_animation_manager method
on the AnimationsDbusServer). When the server attaches an animation
to an AnimationsDbusServerSurface, it uses
the animations_dbus_server_surface_attach_animation_effect_with_server_priority
which gives that attached effect the lowest priority (for instance, if
an external client has already attached a temporary effect to a surface,
that effect has a higher priority than the one attached by the server). This
allows the server to, for instance, globally attach effects to all surfaces
as the "default", but with the ability for external clients on the bus
to override that configuration temporarily.

https://phabricator.endlessm.com/T23314