animations-dbus: More carefully track GTask ownership transfer
Add `g_steal_pointer()` calls in a few places to clarify the ownership
transfer of a reference to the `GTask`. Add a few `g_autoptr()` calls to
ensure that the reference is dropped in all the appropriate places too —
this fixes two refcount leaks on two paths.

Spotted and diagnosed by Daniel Garcia Moreno.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T29887