server-object: Add a stop() method
This will be used to allow the D-Bus service to be stopped (all objects
unexported and the D-Bus well known name unowned) without having to
reach finalisation of the server object.

Tying finalisation of the object to unowning the name makes it difficult
to guarantee the name is unowned when the calling code’s final reference
to the server is dropped, as another bit of code (or the garbage
collector) may still hold a reference.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T29887